backup_manual_code/BU_requirements.txt
connexion == 2.2.0
elasticsearch == 7.6
python_dateutil == 2.6.0
setuptools >= 21.0.0
Flask-Testing >= 0.8.0